requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
ultralytics==8.0.186
opencv-python==4.8.1
pillow==10.1.0
numpy==1.24.0

sqlalchemy==2.0.23
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
